controllers terminés. A tester avec un front. Mais surtout avec un clavier

+++ API_Arcadia/API_Arcadia/Services/AnimalService.cs
@@ -115,8 +115,8 @@ namespace API_Arcadia.Services
                 List<AnimalImage> images = await req.ToListAsync();
                 foreach (var image in images)
                 {
-                    File.Delete(image.Slug);
-                    File.Delete(image.MiniSlug);
+                    if (!String.IsNullOrEmpty(image.Slug)) File.Delete(image.Slug);
+                    if (!String.IsNullOrEmpty(image.MiniSlug)) File.Delete(image.MiniSlug);
                 }
 
                 _context.Animals.Remove(animal);
